Source/eDCMS/requirements.txt
amqp==2.5.0
anyjson==0.3.3
backports.functools-lru-cache==1.5
billiard==3.6.0.0
bootstrap4==0.1.0
celery==4.3.0
cheroot==6.5.6
CherryPy==18.1.2
Django==2.2.3
django-appconf==1.0.3
django-celery-email==2.0.2
django-celery-results==1.1.2
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-froala-editor==3.0.4
django-js-asset==1.2.2
django-report-builder==6.3.1
django-simple-history==2.7.3
django-tinymce==2.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.24.1
greenlet==0.4.15
imutils==0.5.2
jaraco.functools==2.0
jdcal==1.4.1
kombu==4.6.3
monotonic==1.5
more-itertools==7.2.0
mysql-connector-python==8.0.16
numpy==1.17.0
opencv-python==4.1.0.25
openpyxl==2.6.2
Pillow==6.1.0
portend==2.5
protobuf==3.8.0
pypiwin32==223
python-dateutil==2.8.0
pytz==2019.1
pywin32==224
pyzbar==0.1.8
six==1.12.0
sqlparse==0.3.0
tempora==1.14.1
vine==1.3.0
virtualenv==16.6.1
zc.lockfile==2.0
